+++ kotlinx-coroutines-core/common/src/channels/Produce.kt
@@ -26,6 +26,32 @@ public interface ProducerScope<in E> : CoroutineScope, SendChannel<E> {
 }
 
 /**
+ * Suspends the current coroutine until the channel is either [closed][SendChannel.close] or [cancelled][ReceiveChannel.cancel]
+ * and invokes the given [block] before resuming the coroutine.
+ *
+ * Example of usage:
+ * ```
+ * val callbackEventsStream = produce {
+ *     val disposable = registerChannelInCallback(channel)
+ *     await { disposable.dispose() }
+ * }
+ * ```
+ */
+@ExperimentalCoroutinesApi
+public suspend fun <T> ProducerScope<T>.await(block: () -> Unit = {}) {
+    check(kotlin.coroutines.coroutineContext[Job] === this) { "await() can be invoke only from the producer context" }
+    suspendCancellableCoroutine<Unit> { cont ->
+        invokeOnClose {
+            try {
+                block()
+            } finally {
+                cont.resume(Unit)
+            }
+        }
+    }
+}
+
+/**
  * Launches new coroutine to produce a stream of values by sending them to a channel
  * and returns a reference to the coroutine as a [ReceiveChannel]. This resulting
  * object can be used to [receive][ReceiveChannel.receive] elements produced by this coroutine.
